\name{get.value}
\alias{get.value}
\title{Return all tables}
\description{
generic value setting method
}
\usage{
get.value(dlg.item, ...)
}
\arguments{
\item{dlg.item}{gtkWidget}
\item{...}{args}
}
\details{
This function can be called on dialogItems to get the value they are set to.
}
\keyword{interface}
