\name{gtkDfEditCommandData}
\alias{gtkDfEditCommandData}
\title{Open the Command Editor}
\description{
Applies a string command
}
\usage{
gtkDfEditCommandData(x)
}
\arguments{
\item{x}{string x}
}
\details{
This function opens the command editor which allows R commands to be applied to
a selected data submatrix. See ?gtkDfEdit
}
\keyword{interface}
