\name{gtkDfEditGetColumnNames}
\alias{gtkDfEditGetColumnNames}
\title{Return the columns of the RGtk2DfEdit object...}
\usage{gtkDfEditGetColumnNames(object)}
\description{Return the columns of the RGtk2DfEdit object}
\value{Returns the column names for the current object}
\arguments{\item{object}{The RGtk2DfEdit object}}
