\name{gtkDfEditGetDataFrame}
\alias{gtkDfEditGetDataFrame}
\title{Return a data frame from the RGtk2DfEdit object...}
\usage{gtkDfEditGetDataFrame(object)}
\description{Return a data frame from the RGtk2DfEdit object}
\value{Returns the data frame with row names and column names}
\arguments{\item{object}{The RGtk2DfEdit object}}
