\name{gtkDfEditGetDatasetName}
\alias{gtkDfEditGetDatasetName}
\title{Return all tables}
\description{
Gets dataset name
}
\usage{
gtkDfEditGetDatasetName(x)
}
\arguments{
\item{x}{The S3 object}
}
\details{
This method returns the name of the gtkDfEdit dataset.
}
\keyword{interface}
