\name{gtkDfEditGetDimension}
\alias{gtkDfEditGetDimension}
\title{Return the dimensions (nrow, ncol) of the RGtk2DfEdit object...}
\usage{gtkDfEditGetDimension(object)}
\description{Return the dimensions (nrow, ncol) of the RGtk2DfEdit object}
\value{Returns the number of rows and columns -- not counting row names}
\arguments{\item{object}{The RGtk2DfEdit object}}
