\name{gtkDfEditGetRowNames}
\alias{gtkDfEditGetRowNames}
\title{Return the row names of the RGtk2DfEdit object...}
\usage{gtkDfEditGetRowNames(object)}
\description{Return the row names of the RGtk2DfEdit object}
\value{Returns the row names for the current object}
\arguments{\item{object}{The RGtk2DfEdit object}}
