\name{gtkDfEditSetDatasetName}
\alias{gtkDfEditSetDatasetName}
\title{Return all tables}
\description{
Sets name
}
\usage{
gtkDfEditSetDatasetName(x, new.name)
}
\arguments{
\item{x}{x}
\item{new.name}{new.name}
}
\details{
Sets the name of the data set to new.name, in .GlobalEnv
}
\keyword{interface}
