\name{object.exists}
\alias{object.exists}
\title{Return all tables}
\description{
Does object of form a\$b exist
}
\usage{
object.exists(text, envir=parent.env(environment()))
}
\arguments{
\item{text}{text to evaluate}
\item{envir}{environment}
}
\details{
Does object of form a\$b exist
}
\keyword{interface}
