\name{quick_message}
\alias{quick_message}
\title{Return all tables}
\description{
Create a modal message box.
}
\usage{
quick_message(message="", win=NULL)
}
\arguments{
\item{message}{message to use}
\item{win}{gtkWindow}
}
\details{
Create a quick message box.
}
\keyword{interface}
