\name{[.GtkDfEdit}
\alias{[.GtkDfEdit}
\title{S3 data extraction method...}
\usage{\method{[}{GtkDfEdit} (x, i, j, drop=TRUE)}
\description{S3 data extraction method}
\details{Grabs data frame then passes onto [.data.frame method}
\value{The extracted entries}
\arguments{\item{x}{The RGtk2DfEdit object}
\item{i}{Row index}
\item{j}{Column index}
\item{drop}{passed to extraction for data frame}}



