\name{my_choose_files}
\alias{my_choose_files}
\title{Return all tables}
\description{
Platform independent file chooser
}
\usage{
my_choose_files(fn = "*.*", multi=FALSE, filters=NULL, type = "open")
}
\arguments{
\item{fn}{filename}
\item{multi}{multiple}
\item{filters}{Win filters}
\item{type}{type of dialog}
}
\details{
}
\keyword{interface}
